# GEODE-6404: work around possible sync issue with computeIfAbsent