@pivotal-jbarrett 2019-02-17
Given that Java 9+ has fixed this I’d suggest a check here for the JVM version and disable if 9+.